Add sub components for UI
