+++ plugins/mta/src/components/MTAApplicationManager/MTAApplicationManager.tsx
@@ -13,68 +13,69 @@ import { useApi } from '@backstage/core-plugin-api';
 import { mtaApiRef, Application, MTAApi } from '../../api/api';
 import AddLinkIcon from '@mui/icons-material/AddLink';
 import { useEntity } from '@backstage/plugin-catalog-react';
+import { DenseApplicationTable } from '../DenseApplicationTable/DenseApplicationTable';
 
-type DenseApplicationTableProps = {
-  applications: Application[];
-  api: MTAApi;
-  entityID: string;
-};
+// type DenseApplicationTableProps = {
+//   applications: Application[];
+//   api: MTAApi;
+//   entityID: string;
+// };
 
-export const DenseApplicationTable = ({
-  applications,
-  api,
-  entityID,
-}: DenseApplicationTableProps) => {
-  const columns: TableColumn[] = [
-    { title: 'Name', field: 'name' },
-    { title: 'Description', field: 'description' },
-    { title: 'Assessed', field: 'assessed' },
-    { title: 'mtaID', field: 'mtaID', hidden: true },
-  ];
+// export const DenseApplicationTable = ({
+//   applications,
+//   api,
+//   entityID,
+// }: DenseApplicationTableProps) => {
+//   const columns: TableColumn[] = [
+//     { title: 'Name', field: 'name' },
+//     { title: 'Description', field: 'description' },
+//     { title: 'Assessed', field: 'assessed' },
+//     { title: 'mtaID', field: 'mtaID', hidden: true },
+//   ];
 
-  const data = applications.map(application => {
-    return {
-      name: application.name,
-      description: application.description,
-      assessed: application.assessed,
-      mtaID: application.id,
-    };
-  });
+//   const data = applications.map(application => {
+//     return {
+//       name: application.name,
+//       description: application.description,
+//       assessed: application.assessed,
+//       mtaID: application.id,
+//     };
+//   });
 
-  return (
-    // add Info cards here, that explain to attach an application to the enity.
-    <>
-      <InfoCard title="Link your Application to Component">
-        <Table
-          title="Application List"
-          options={{ search: false, paging: true }}
-          columns={columns}
-          data={data}
-          actions={[
-            {
-              icon: () => (
-                <>
-                  <AddLinkIcon />
-                </>
-              ),
-              tooltip: 'Connect Application to MTA Application',
-              onClick: (event, rowData: any) => {
-                // console.log(rowData);
-                // console.log(event);
-                if (Array.isArray(rowData)) {
-                  // console.log('unable to handle array of row data');
-                } else {
-                  const { mtaID } = rowData;
-                  const res = api.saveApplicationEntity(mtaID, entityID);
-                }
-              },
-            },
-          ]}
-        />
-      </InfoCard>
-    </>
-  );
-};
+//   return (
+//     // add Info cards here, that explain to attach an application to the enity.
+//     <>
+//       <InfoCard title="Link your Application to Component">
+//         <Table
+//           title="Application List"
+//           options={{ search: false, paging: true }}
+//           columns={columns}
+//           data={data}
+//           actions={[
+//             {
+//               icon: () => (
+//                 <>
+//                   <AddLinkIcon />
+//                 </>
+//               ),
+//               tooltip: 'Connect Application to MTA Application',
+//               onClick: (event, rowData: any) => {
+//                 // console.log(rowData);
+//                 // console.log(event);
+//                 if (Array.isArray(rowData)) {
+//                   // console.log('unable to handle array of row data');
+//                 } else {
+//                   const { mtaID } = rowData;
+//                   const res = api.saveApplicationEntity(mtaID, entityID);
+//                 }
+//               },
+//             },
+//           ]}
+//         />
+//       </InfoCard>
+//     </>
+//   );
+// };
 
 type loginPageProps = {
   url: URL;
@@ -108,7 +109,7 @@ export const ApplicationCard = ({ application }: ApplicationCardProps) => {
   );
 };
 
-export const ExampleFetchComponent = () => {
+export const MTAApplicationManager = () => {
   const api = useApi(mtaApiRef);
   const entity = useEntity();
   console.log('Entity data:', entity);
